fix project

+++ app/src/main/java/hexlet/code/schemas/NumberSchema.java
@@ -5,13 +5,13 @@ import java.util.function.Predicate;
 public class NumberSchema extends BaseSchema<Integer> {
 
     public final NumberSchema positive() {
-        Predicate<Integer> positive = s -> s == null || (int) s > 0;
+        Predicate<Integer> positive = s -> s == null || s > 0;
         addCheck("positive", positive);
         return this;
     }
 
     public final NumberSchema range(int a, int b) {
-        Predicate<Integer> range = s -> s  == null || (int) s >= a && (int) s <= b;
+        Predicate<Integer> range = s -> s  == null || s >= a && s <= b;
         addCheck("range", range);
         return this;
     }
